requirements.txt
streamlit
pdfplumber
python-docx
spacy
langdetect
scikit-learn
reportlab
